refactor: Money 클래스 패키지 위치 변경
