+++ src/test/java/shopping/cart/service/CartServiceTest.java
@@ -17,7 +17,7 @@ import shopping.cart.repository.CartItemRepository;
 import shopping.member.domain.Member;
 import shopping.member.repository.MemberRepository;
 import shopping.product.domain.Product;
-import shopping.product.domain.vo.Money;
+import shopping.common.vo.Money;
 import shopping.product.repository.ProductRepository;
 
 @DataJpaTest
